requirements.txt
# Core Framework
streamlit==1.28.2
streamlit-authenticator==0.2.3
streamlit-extras==0.3.6

# Streamlit Components
streamlit-aggrid==0.3.4.post3
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-embedcode==0.1.2
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-option-menu==0.3.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
extra-streamlit-components==0.1.60
# st-annotated-text==4.0.0  # Commented out - has dependency conflict and not used in codebase

# HTTP & API Clients
requests==2.31.0
aiohttp==3.9.5
httpx==0.24.1
httpcore==0.17.2
urllib3==1.26.14
certifi==2022.12.7
folioclient>=0.61.0

# Async & Concurrency
aiofiles>=25.0.0
anyio==3.7.0
async-generator==1.10
async-timeout==4.0.3
aiosignal==1.3.1
frozenlist==1.4.1
multidict==6.0.5
yarl==1.9.4
h11==0.14.0
outcome==1.2.0
sniffio==1.3.0
trio==0.22.0
trio-websocket==0.10.2
wsproto==1.2.0

# Data Processing
pandas==2.1.3
numpy==1.24.2
openpyxl==3.1.2
pyarrow==11.0.0
et-xmlfile==1.1.0

# Web Scraping & Parsing
beautifulsoup4==4.12.2
soupsieve==2.4.1
lxml==4.9.2
pymarc==5.1.2

# Authentication & Security
bcrypt==4.0.1
PyJWT==2.6.0
PySocks==1.7.1
cffi==1.15.1
pycparser==2.21

# Browser Automation
selenium==4.8.3

# Data Validation & Serialization
PyYAML==6.0.1
jsonschema==4.17.3
pyrsistent==0.19.3
validators==0.20.0

# Date & Time
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
tzdata==2022.7
tzlocal==4.2

# Utilities
click==8.1.3
colorama==0.4.6
python-decouple==3.7
typer>=0.20.0
rich==13.3.1
tenacity==8.2.2
semver==2.13.0
Faker==18.4.0
favicon==0.7.0

# Markdown & Documentation
Markdown==3.4.3
markdown-it-py==2.1.0
markdownlit==0.0.6
mdurl==0.1.2
pymdown-extensions==9.11
Pygments==2.14.0

# Visualization
altair==4.2.2
matplotlib==3.7.1
pydeck==0.8.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
kiwisolver==1.4.4
pyparsing==3.0.9
Pillow==9.4.0

# Web Framework
flask==2.2.2
Jinja2==3.1.2
MarkupSafe==2.1.2
blinker==1.5

# System & OS
watchdog==2.2.1
GitPython==3.1.31
gitdb==4.0.10
smmap==5.0.0

# Configuration & Metadata
toml==0.10.2
importlib-metadata==6.0.0
importlib-resources==5.12.0
entrypoints==0.4
packaging==23.0

# Type Checking & Extensions
typing_extensions==4.5.0
attrs==22.2.0
exceptiongroup==1.1.1

# Other Dependencies
cachetools==5.3.0
charset-normalizer==3.0.1
decorator==5.1.1
idna==3.4
more-itertools==9.1.0
protobuf==3.20.3
prometheus_client==0.20.0
Pympler==1.0.1
sortedcontainers==2.4.0
six==1.16.0
toolz==0.12.0
tornado==6.2
zipp==3.14.0
htbuilder==0.6.2
